web/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pillow
pyautogui
requests
google-auth-oauthlib
google-api-python-client
-e ../..  # Install the root project (hyperpocket) in editable mode 